Introduce flag to control (non-)blocking wait for CUDAFuture
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50047

Waiting on a CUDAFuture consists of two steps: first we wait for the result to be populated (just like with the ordinary CPU future) and then we wait for any async kernel on that result to complete. This second step can be done in two ways:
- a non-blocking way, where the CPU thread returns immediately, after it simply adds the necessary instructions on the CUDA streams to ensure that future work on the result will be correctly scheduled.
- a blocking way, where the CPU thread is stalled until the kernels effectively complete on the device.

We want to support both modes, since the non-blocking one is more idiomatic (for CUDA) and more performant (thus power-users will want it), whereas the blocking one is the more beginner-friendly.

We'll allow to control between the two modes with a flag on the `wait` method. We need to add this flag also to the CPU-only ivalue::Future (which is the parent class of CUDAFuture), even though it will have no effect there, since CUDAFuture is a "private" class and is only returned as a down-cast pointer to ivalue::Future, hence the flag needs to be available on the latter too.

The behavior of CUDAFuture so far was non-blocking, but we want the default to be blocking (because it's safer), hence this commit introduces a functional change. Note that we don't expect breakages, since a blocking wait should behave the same as a non-blocking wait, but we might see performance losses.

At first I thought that we could "code-mod" all current usages of `wait` in PyTorch's codebase to set the non-blocking flag to true, because those callsites were already operating in non-blocking mode prior to this change. However, (non-)blockingness only matters for CUDA futures and those were until recently only used by ProcessGroupNCCL, and thus those callsites might never have been tested with a CUDAFuture and therefore we can't be sure that they can deal with non-blockingness. Hence to be safe I let them become blocking, and I added a FIXME. If the reviewers can suggest a POC for those callsites we can ask them whether it's safe to use a non-blocking wait.

The only place I could track down where CUDA futures were certainly used was the DDP comm hook in the reducer, hence that's the only place where I explicitly set the non-blocking flag to true.
ghstack-source-id: 119375964

Differential Revision: [D25744506](https://our.internmc.facebook.com/intern/diff/D25744506/)