Push anonymous namespace into codegen, not template
In the near future, I want to code generate some functions that are
visible externally to this compilation unit.  I cannot easily do this
if all the codegen code is wrapped in a global anonymous namespace,
so push the namespace in.

Registration has to stay in an anonymous namespace to avoid name conflicts.
This could also have been solved by making the wrapper functions have
more unique names but I didn't do this in the end.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 3f7bdd7d9eb45aec87a4445adb610c82ef377651
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49498